Move callback code to `common`.
The 4 common callback routines are moved to the common folder as they're
identical between the examples.
