+++ src/example_01/CMakeLists.txt
@@ -15,6 +15,8 @@
 add_executable(dusk_example_01)
 dusk_compile_options(dusk_example_01)
 target_sources(dusk_example_01 PRIVATE
+  ../common/callback.cc
+  ../common/callback.h
   ../common/log.cc
   ../common/log.h
   ../common/wgpu.h
