+++ src/example_05/CMakeLists.txt
@@ -15,6 +15,8 @@
 add_executable(dusk_example_05 ${DUSK_SRCS})
 dusk_compile_options(dusk_example_05)
 target_sources(dusk_example_05 PRIVATE
+  ../common/callback.cc
+  ../common/callback.h
   ../common/glfw.h
   ../common/log.cc
   ../common/log.h
